refactor: rename storage utility function and enhance image deletion …

+++ django_ckeditor_5/signals.py
@@ -5,7 +5,7 @@ from django.db.models.signals import pre_delete, pre_save
 from django.dispatch import receiver
 
 from django_ckeditor_5.fields import CKEditor5Field
-from django_ckeditor_5.storage_utils import get_storage_class
+from django_ckeditor_5.storage_utils import get_django_storage_class
 
 import logging
 
@@ -30,11 +30,26 @@ def delete_images(storage, image_paths):
     """
     Deletes images from disk based on the provided paths.
     """
-    for img_path in image_paths:
-        file_name = os.path.basename(img_path)
-        abs_path = os.path.join(storage.location, file_name)
-        if os.path.exists(abs_path):
-            os.remove(abs_path)
+    if hasattr(storage, "location"):
+        for img_path in image_paths:
+            file_name = os.path.basename(img_path)
+            abs_path = os.path.join(storage.location, file_name)
+            if os.path.exists(abs_path):
+                os.remove(abs_path)
+    else:
+        logger.warning(
+            "Storage does not have 'location' attribute, skipping file deletion."
+        )
+
+
+def get_safe_storage():
+    """Safely returns a storage instance based on Django settings."""
+    try:
+        storage_class = get_django_storage_class()
+        return storage_class()
+    except Exception as e:
+        logger.error(f"Could not initialize storage class: {e}")
+        return None
 
 
 @receiver(pre_delete)
@@ -43,10 +58,10 @@ def cleanup_ckeditor_images_on_delete(sender, instance, **kwargs):
     Removes images from disk when an object is deleted.
     If an error occurs, it is logged, but the deletion process continues.
     """
+    if not any(isinstance(f, CKEditor5Field) for f in instance._meta.fields):
+        return
     try:
-        storage_class = get_storage_class()
-        storage = storage_class()
-
+        storage = get_safe_storage()
         if not storage:
             return
 
@@ -72,8 +87,10 @@ def cleanup_unused_ckeditor_images_on_update(sender, instance, **kwargs):
     Removes unused images when an object is updated.
     If any unexpected error occurs, it will be logged, but the deletion process won't break the update.
     """
+    if not any(isinstance(f, CKEditor5Field) for f in instance._meta.fields):
+        return
     try:
-        storage = get_storage_class()
+        storage = get_safe_storage()
         if not storage:
             return
